Move resizing hook into its own file
